🐛 fix: say every rate in bytes, which is what the engine counts
Every speed in the app read eight times what it was. Bytes per second
were converted to megabits and then labelled "MB/s", so the 5.7 MB/s
peak of a real download showed as 45.5 MB/s — and the size beside it,
in real megabytes, could not be compared with it without knowing that.

There is one unit now, end to end: bytes per second, the same multiples
a file size uses, so a rate beside a size means what it looks like it
means. `formatRate` scales through B/s, KB/s, MB/s and GB/s exactly as
`formatBytes` scales, and the rate cap in settings goes through it
rather than carrying its own half of the same logic.

The thresholds that were tuned in megabits are restated in bytes rather
than reconverted: the glow saturates at a megabyte a second and turns
vivid at half of one, which are the speeds the old numbers described
said in the unit that is actually counted.

+++ portalis/lib/shell/desktop/layout.dart
@@ -20,7 +20,7 @@ class DesktopShellLayout extends StatelessWidget {
   final ValueChanged<DesktopPane> onPane;
   final Widget home;
   final Widget content;
-  final double liveRate;
+  final int liveRate;
 
   @override
   Widget build(BuildContext context) => Scaffold(
